# app/swagger_spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Doc Manager API",
    "description": "Swagger/OpenAPI 문서의 버전 관리 및 변경사항 추적 시스템 API",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "API 서버"
    }
  ],
...
  "paths": {
...
    "/urls": {
      "get": {
        "tags": [
          "URLs"
        ],
        "summary": "URL 목록 조회",
        "description": "등록된 API URL 목록을 조회합니다. 필터링, 검색, 정렬, 페이지네이션을 지원합니다.\n\n**사용 예시:**\n- 모든 API 조회: `GET /api/urls`\n- 팀별 필터링: `GET /api/urls?group=backend`\n- 검색: `GET /api/urls?search=결제`\n- 정렬: `GET /api/urls?sort=group service` (팀별, 서비스별 정렬)\n- 페이지네이션: `GET /api/urls?page=2&limit=20`",
        "operationId": "getUrls",
        "parameters": [
          {
            "name": "group",
            "in": "query",
            "description": "그룹(팀)별 필터링. 소문자로 자동 변환됩니다.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "backend"
            }
          },
          {
            "name": "service",
            "in": "query",
            "description": "서비스별 필터링. 소문자로 자동 변환됩니다.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "user-service"
            }
          },
          {
            "name": "isActive",
            "in": "query",
            "description": "활성화 상태 필터. true 또는 false 문자열로 전달합니다.",
            "required": false,
            "schema": {
              "type": "boolean",
              "example": true
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "이름(name) 또는 설명(description) 필드에서 텍스트 검색. 대소문자 구분 없음.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "결제"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "정렬 기준. 공백으로 구분하며, `-` 접두사로 내림차순 지정 가능.\n예: `-updatedAt` (최신순), `group service` (팀별, 서비스별 오름차순)",
            "required": false,
            "schema": {
              "type": "string",
              "default": "-updatedAt",
              "example": "-updatedAt"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "페이지 번호 (1부터 시작)",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1,
              "example": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "페이지당 항목 수. 최대 100개까지 가능.",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 100,
              "example": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UrlListResponse"
                },
                "example": {
                  "success": true,
                  "data": [
                    {
                      "_id": "507f1f77bcf86cd799439011",
                      "name": "User Service API",
                      "url": "https://api.example.com/swagger.json",
                      "group": "backend",
                      "service": "user-service",
                      "description": "사용자 관리 API",
                      "isActive": true,
                      "versionCount": 5,
                      "lastFetchedAt": "2024-11-27T10:00:00.000Z",
                      "lastFetchStatus": "success",
                      "createdAt": "2024-11-20T10:00:00.000Z",
                      "updatedAt": "2024-11-27T10:00:00.000Z"
                    }
                  ],
                  "meta": {
                    "total": 25,
                    "page": 1,
                    "limit": 50,
                    "totalPages": 1,
                    "groups": [
                      "backend",
                      "frontend",
                      "platform"
                    ],
                    "services": [
                      "user-service",
                      "auth-service"
                    ],
                    "servicesByGroup": {
                      "backend": [
                        "user-service",
                        "auth-service"
                      ],
                      "frontend": [
                        "web-app"
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "429": {
            "description": "요청 제한 초과",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": {
                    "code": "RATE_LIMIT_EXCEEDED",
                    "message": "요청이 너무 많습니다. 잠시 후 다시 시도해주세요."
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": [
          "URLs"
        ],
        "summary": "URL 등록",
        "description": "새 API URL을 등록합니다. Swagger/OpenAPI JSON을 반환하는 URL을 등록하면 자동으로 버전 관리가 시작됩니다.\n\n**필수 필드:**\n- `name`: API 이름 (최대 100자)\n- `url`: Swagger JSON URL (HTTP/HTTPS)\n- `group`: 팀/그룹명\n- `service`: 서비스명\n\n**선택 필드:**\n- `description`: 설명 (최대 500자)\n- `owner`: 담당자 이메일\n- `tags`: 태그 배열\n- `priority`: 우선순위 (low/medium/high)",
        "operationId": "createUrl",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUrlRequest"
              },
              "example": {
                "name": "User Service API",
                "url": "https://api.example.com/v3/api-docs",
                "group": "backend",
                "service": "user-service",
                "description": "사용자 관리 API",
                "owner": "developer@example.com",
                "tags": [
                  "v2",
                  "production"
                ],
                "priority": "high"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "생성 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UrlResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "_id": "507f1f77bcf86cd799439011",
                    "name": "User Service API",
                    "url": "https://api.example.com/v3/api-docs",
                    "group": "backend",
                    "service": "user-service",
                    "description": "사용자 관리 API",
                    "isActive": true,
                    "createdAt": "2024-11-27T10:00:00.000Z",
                    "updatedAt": "2024-11-27T10:00:00.000Z"
                  },
                  "message": "URL이 성공적으로 추가되었습니다"
                }
              }
            }
          },
          "400": {
            "description": "입력값 검증 실패 또는 중복 URL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "validationError": {
                    "summary": "유효성 검사 실패",
                    "value": {
                      "success": false,
                      "error": {
                        "code": "VALIDATION_ERROR",
                        "message": "입력값 검증 실패",
                        "details": [
                          "서비스명은 필수입니다",
                          "URL은 필수입니다",
                          "유효한 HTTP(S) URL이어야 합니다"
                        ]
                      }
                    }
                  },
                  "duplicateUrl": {
                    "summary": "중복 URL",
                    "value": {
                      "success": false,
                      "error": {
                        "code": "DUPLICATE_URL",
                        "message": "이미 존재하는 url입니다"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/urls/{id}": {
      "get": {
        "tags": [
          "URLs"
        ],
        "summary": "URL 상세 조회",
        "description": "특정 API URL의 상세 정보를 조회합니다. 버전 수, 최종 업데이트 시간, 상태 등의 정보를 포함합니다.",
        "operationId": "getUrlById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "API URL의 MongoDB ObjectId (24자리 hex 문자열)",
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{24}$",
              "example": "507f1f77bcf86cd799439011"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UrlResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "_id": "507f1f77bcf86cd799439011",
                    "name": "User Service API",
                    "url": "https://api.example.com/swagger.json",
                    "group": "backend",
                    "service": "user-service",
                    "description": "사용자 관리 API",
                    "isActive": true,
                    "owner": "developer@example.com",
                    "tags": [
                      "v2",
                      "production"
                    ],
                    "priority": "high",
                    "versionCount": 5,
                    "lastFetchedAt": "2024-11-27T10:00:00.000Z",
                    "lastFetchStatus": "success",
                    "createdAt": "2024-11-20T10:00:00.000Z",
                    "updatedAt": "2024-11-27T10:00:00.000Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 ID 형식",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": {
                    "code": "INVALID_ID",
                    "message": "잘못된 ID 형식입니다"
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "put": {
        "tags": [
          "URLs"
        ],
        "summary": "URL 수정",
        "description": "API URL 정보를 수정합니다. 수정할 필드만 전송하면 됩니다. 모든 필드는 선택사항이지만 최소 1개 필드는 수정해야 합니다.",
        "operationId": "updateUrl",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "API URL의 MongoDB ObjectId",
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{24}$",
              "example": "507f1f77bcf86cd799439011"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUrlRequest"
              },
              "example": {
                "name": "Updated API Name",
                "description": "Updated description",
                "priority": "medium"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "수정 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UrlResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "_id": "507f1f77bcf86cd799439011",
                    "name": "Updated API Name",
                    "url": "https://api.example.com/swagger.json",
                    "group": "backend",
                    "service": "user-service",
                    "description": "Updated description",
                    "priority": "medium",
                    "updatedAt": "2024-11-27T11:00:00.000Z"
                  },
                  "message": "URL이 성공적으로 수정되었습니다"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "tags": [
          "URLs"
        ],
        "summary": "URL 삭제",
        "description": "API URL을 삭제합니다. **주의:** 삭제 시 모든 버전 히스토리도 함께 삭제되며, 이 작업은 되돌릴 수 없습니다.",
        "operationId": "deleteUrl",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "API URL의 MongoDB ObjectId",
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{24}$",
              "example": "507f1f77bcf86cd799439011"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "삭제 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessageResponse"
                },
                "example": {
                  "success": true,
                  "message": "URL이 성공적으로 삭제되었습니다"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/urls/{id}/activate": {
      "patch": {
        "tags": [
          "URLs"
        ],
        "summary": "URL 활성화/비활성화",
        "description": "API URL의 활성화 상태를 토글합니다. 현재 상태가 활성화(`isActive: true`)이면 비활성화로, 비활성화(`isActive: false`)이면 활성화로 변경됩니다. 비활성화된 API는 목록에서 필터링할 수 있습니다.",
        "operationId": "toggleUrlActivation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "API URL의 MongoDB ObjectId",
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{24}$",
              "example": "507f1f77bcf86cd799439011"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivationResponse"
                },
                "examples": {
                  "activated": {
                    "summary": "활성화됨",
                    "value": {
                      "success": true,
                      "data": {
                        "isActive": true
                      },
                      "message": "URL이 활성화되었습니다"
                    }
                  },
                  "deactivated": {
                    "summary": "비활성화됨",
                    "value": {
                      "success": true,
                      "data": {
                        "isActive": false
                      },
                      "message": "URL이 비활성화되었습니다"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/urls/{id}/fetch": {
      "post": {
        "tags": [
          "URLs"
        ],
        "summary": "Swagger JSON 가져오기",
        "description": "Swagger JSON을 수동으로 가져와 버전을 생성/업데이트합니다. 등록된 URL에서 Swagger JSON을 다운로드하고, 이전 버전과 비교하여 변경사항을 분석합니다. 변경사항이 있으면 새 버전을 생성하고, 없으면 기존 버전을 유지합니다.\n\n**주의사항:**\n- 네트워크 요청이므로 시간이 걸릴 수 있습니다.\n- Swagger URL이 접근 불가능하거나 잘못된 형식이면 에러가 발생합니다.",
        "operationId": "fetchSwagger",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "API URL의 MongoDB ObjectId",
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{24}$",
              "example": "507f1f77bcf86cd799439011"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공 (새 버전 생성 또는 변경사항 없음)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FetchResponse"
                },
                "examples": {
                  "newVersionCreated": {
                    "summary": "새 버전 생성됨",
                    "value": {
                      "success": true,
                      "data": {
                        "created": true,
                        "updated": false,
                        "version": {
                          "_id": "507f1f77bcf86cd799439013",
                          "versionId": "v4",
                          "versionNumber": 4,
                          "endpointCount": 16,
                          "summary": "1개 추가, 0개 삭제, 2개 수정"
                        }
                      },
                      "message": "새 버전이 생성되었습니다"
                    }
                  },
                  "noChanges": {
                    "summary": "변경사항 없음",
                    "value": {
                      "success": true,
                      "data": {
                        "created": false,
                        "updated": false,
                        "version": null
                      },
                      "message": "변경사항이 없습니다"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Swagger JSON 다운로드 실패 또는 잘못된 형식",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "fetchError": {
                    "summary": "다운로드 실패",
                    "value": {
                      "success": false,
                      "error": {
                        "code": "FETCH_ERROR",
                        "message": "Swagger JSON을 가져올 수 없습니다: Network timeout"
                      }
                    }
                  },
                  "invalidSwagger": {
                    "summary": "잘못된 Swagger 형식",
                    "value": {
                      "success": false,
                      "error": {
                        "code": "INVALID_SWAGGER",
                        "message": "유효하지 않은 Swagger/OpenAPI 형식입니다"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UrlResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "$ref": "#/components/schemas/ApiUrl"
          },
          "message": {
            "type": "string",
            "example": "URL이 성공적으로 추가되었습니다"
          }
        }
      },
      "UrlListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiUrl"
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "total": {
                "type": "integer",
                "example": 25
              },
              "page": {
                "type": "integer",
                "example": 1
              },
              "limit": {
                "type": "integer",
                "example": 10
              },
              "totalPages": {
                "type": "integer",
                "example": 3
              },
              "groups": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": [
                  "backend",
                  "frontend"
                ]
              },
              "services": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": [
                  "user-service",
                  "auth-service"
                ]
              },
              "servicesByGroup": {
                "type": "object",
                "additionalProperties": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "example": {
                  "backend": [
                    "user-service",
                    "auth-service"
                  ]
                }
              }
            }
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "잘못된 요청 (유효성 검사 실패 또는 잘못된 파라미터)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "validationError": {
                "summary": "입력값 검증 실패",
                "value": {
                  "success": false,
                  "error": {
                    "code": "VALIDATION_ERROR",
                    "message": "입력값 검증 실패",
                    "details": [
                      "서비스명은 필수입니다",
                      "URL은 필수입니다",
                      "유효한 HTTP(S) URL이어야 합니다"
                    ]
                  }
                }
              },
              "invalidId": {
                "summary": "잘못된 ID 형식",
                "value": {
                  "success": false,
                  "error": {
                    "code": "INVALID_ID",
                    "message": "잘못된 ID 형식입니다"
                  }
                }
              },
              "duplicateUrl": {
                "summary": "중복 URL",
                "value": {
                  "success": false,
                  "error": {
                    "code": "DUPLICATE_URL",
                    "message": "이미 존재하는 url입니다"
                  }
                }
              }
            }
          }
        }
      },
...
    }
  }
}